+++ src/components/Home/Tags/Tags.module.css
@@ -1,0 +1,15 @@
+.container {
+    display: flex;
+    flex-direction: column;
+    width: 268px;
+    height: 124px;
+    background: var(--background-secondary);
+    border-radius: 20px;
+    padding: 16px;
+}
+
+:local .title {
+    composes: text from global;
+    color: var(--text-primary-color);
+    margin: 0;
+}
