+++ src/components/common/TagList/TagList.module.css
@@ -1,18 +1,3 @@
-.container {
-    display: flex;
-    flex-direction: column;
-    width: 268px;
-    height: 124px;
-    background: var(--background-secondary);
-    border-radius: 20px;
-    padding: 16px;
-}
-
-:local .title {
-    composes: text from global;
-    color: var(--text-primary-color);
-}
-
 .tags {
     width: 100%;
     display: flex;
